refactor: Simplify temporal pooling integration tests and enhance utility functions
- Refactored the integration tests in `test_temporal_pooling_integration.cpp` to utilize utility functions from `temporal_pooling_test_utils`, improving code readability and maintainability.
- Updated test cases to ensure consistent usage of the new `runSteps` function, which now accepts time steps and input handling as parameters.
- Enhanced the structure of tests to better reflect the intent of measuring temporal pooling across different input patterns and scenarios, including repeating sequences and multi-pattern differentiation.
- Ensured that the tests continue to validate the stability of pooled representations in both active columns and learning cells, aligning with previous Python implementations.
